+++ src/model.js
@@ -37,7 +37,7 @@ import Promise from './base/promise';
  * {@link Model#constructor constructor}, which allows you to replace the
  * actual constructor function for your model.
  *
- *     let Books = bookshelf.Model.extend({
+ *     let Book = bookshelf.Model.extend({
  *       tableName: 'documents',
  *       constructor: function() {
  *         bookshelf.Model.apply(this, arguments);
@@ -79,12 +79,12 @@ const BookshelfModel = ModelBase.extend({
    *
    *     // select * from `health_records` where `patient_id` = 1;
    *     new Patient({id: 1}).related('record').fetch().then(function(model) {
-   *       ...
+   *       // ...
    *     });
    *
    *     // alternatively, if you don't need the relation loaded on the patient's relations hash:
    *     new Patient({id: 1}).record().fetch().then(function(model) {
-   *       ...
+   *       // ...
    *     });
    *
    * @method Model#hasOne
@@ -443,7 +443,6 @@ const BookshelfModel = ModelBase.extend({
    *
    *
    *     let Book = bookshelf.Model.extend({
-   *
    *       tableName: 'books',
    *
    *       // Find all paragraphs associated with this book, by
@@ -455,21 +454,17 @@ const BookshelfModel = ModelBase.extend({
    *       chapters: function() {
    *         return this.hasMany(Chapter);
    *       }
-   *
    *     });
    *
    *     let Chapter = bookshelf.Model.extend({
-   *
    *       tableName: 'chapters',
    *
    *       paragraphs: function() {
    *         return this.hasMany(Paragraph);
    *       }
-   *
    *     });
    *
    *     let Paragraph = bookshelf.Model.extend({
-   *
    *       tableName: 'paragraphs',
    *
    *       chapter: function() {
@@ -480,7 +475,6 @@ const BookshelfModel = ModelBase.extend({
    *       book: function() {
    *         return this.belongsTo(Book).through(Chapter);
    *       }
-   *
    *     });
    *
    * The "through" table creates a pivot model, which it assigns to {@link
@@ -599,7 +593,7 @@ const BookshelfModel = ModelBase.extend({
    *   Reject the returned response with a {@link Model.NotFoundError
    *   NotFoundError} if no results are empty.
    * @param {string|string[]} [options.columns='*']
-   *   Specify columns to be retireved.
+   *   Specify columns to be retrieved.
    * @param {Transaction} [options.transacting]
    *  Optionally run the query in a transaction.
    * @param {string|Object|mixed[]} [options.withRelated]
@@ -1090,7 +1084,6 @@ const BookshelfModel = ModelBase.extend({
        *
        * @event Model#destroying
        * @param {Model}  model    The model firing the event.
-       * @param {Object} attrs    Model firing the event.
        * @param {Object} options  Options object passed to {@link Model#save save}.
        * @returns {Promise}
        */
@@ -1160,10 +1153,14 @@ const BookshelfModel = ModelBase.extend({
    * model.query(function(qb) {
    *   qb.where('other_person', 'LIKE', '%Demo').orWhere('other_id', '>', 10);
    * }).fetch()
-   *   .then(function(model) { // ...
+   *   .then(function(model) {
+   *     // ...
+   *   });
    *
    * let qb = model.query();
-   *     qb.where({id: 1}).select().then(function(resp) { // ...
+   * qb.where({id: 1}).select().then(function(resp) {
+   *   // ...
+   * });
    *
    * @method Model#query
    * @param {function|Object|...string=} arguments The query method.
@@ -1204,6 +1201,34 @@ const BookshelfModel = ModelBase.extend({
    */
   where(...args) {
     return this.query('where', ...args);
+  },
+
+  /**
+   * @method Model#orderBy
+   * @since 0.9.3
+   * @description
+   *
+   * Specifies the column to sort on and sort order.
+   *
+   * The order parameter is optional, and defaults to 'ASC'. You may
+   * also specify 'DESC' order by prepending a hyphen to the sort column
+   * name. `orderBy("date", 'DESC')` is the same as `orderBy("-date")`.
+   *
+   * Unless specified using dot notation (i.e., "table.column"), the default
+   * table will be the table name of the model `orderBy` was called on.
+   *
+   * @example
+   *
+   * Car.forge().orderBy('color', 'ASC').fetchAll()
+   *    .then(function (rows) { // ...
+   *
+   * @param sort {string}
+   *   Column to sort on
+   * @param order {string}
+   *   Ascending ('ASC') or descending ('DESC') order
+   */
+  orderBy(...args) {
+    return Helpers.orderBy(this, ...args);
   },
 
   /* Ensure that QueryBuilder is copied on clone. */